+++ Simplic.CDN.ODBC/Simplic.CDN.ODBC/QueryResult.h
@@ -10,7 +10,7 @@ class ColumnDescriptor
 private:
 	std::string m_name;
 	int16_t m_type;
-	uint64_t m_size;
+	size_t m_size;
 
 	// InfoRecord that is used when the ODBC application requests information about this object.
 	// The InfoRecord object will be initialized on the first request from the application.
@@ -28,7 +28,7 @@ public:
 
 	inline int16_t getType() const { return m_type; } 
 	inline std::string getName() const { return m_name; }
-	inline uint64_t getSize() const { return m_size; }
+	inline size_t getSize() const { return m_size; }
 
 	// used by SQLColAttribute to retrieve fields
 	bool odbcGetField(
